chore: fixed `api` package project structure
