+++ api/match/repository.go
@@ -1,0 +1,8 @@
+/* repository.go
+ * Contains the logic for interacting with the database (MongoDB)
+ * Authors: Zachary Bower
+ * Last modified: 28/05/2025
+ */
+
+package match
+
