+++ api/processing/processor.go
@@ -1,0 +1,7 @@
+/* processor.go
+ * Contains the logic for processing user data. This includes processing results, calulating scores and leaderboards
+ * Authors: Zachary Bower
+ * Last modified: 28/05/2025
+ */
+
+package processor
